refactor(slice-32c): split mongo_stats out of the Mongo storage adapter
Extract db-stats / explore / vector-db aggregation into server/db/mongo_stats.py
so MongoStorageBackend stays under the craft class-size ceiling (Slice 32C M4).
Hoist the core-retriever search import to module scope (no-inline-imports rule).
Behaviour on the Mongo path is unchanged; acceptance tests patch the new module.
